API

EasyLD.download_gnomad_LD_matrices — Function

download_gnomad_LD_matrices(population::String, outdir::String, [start_from], [num_files])

Downloads the LD matrices from gnomAD panel for specified population and saves result to outdir. See https://gnomad.broadinstitute.org/downloads#v2-linkage-disequilibrium.

Inputs

population: Options include

afr for African/African American population

amr for Latino/Admixed American population

asj for Ashkenazi Jewish population

eas for East Asian population

fin for European (Finnish) population

nfe for European (non-Finnish) population

est for Estonian population

nwe for North-western European population

seu for Southern European population

outdir: Directory for which files will be saved.

start_from: Int specifying which file to start downloading at

num_files: Int specifying how many files to download

Multiple download sessions at once

Since there are lots of files to download, one can specify the number of files to download by modifying the num_files files option. This function will download the first num_files file counting from the start_from element of the output of get_gnomad_filenames(). For example, if myfiles = get_gnomad_filenames("afr"), then the function will download files from myfiles[start_from] through myfiles[start_from+num_files-1]

EasyLD.download_ukb_LD_matrices — Function

download_ukb_LD_matrices(population::String, outdir::String, [start_from], [num_files])

Downloads the LD matrices from pan-UK-biobank for specified population and saves result to outdir. See https://pan-dev.ukbb.broadinstitute.org/docs/hail-format/index.html.

Inputs

population: Several options are available. These codes refer only to ancestry groupings used in GWAS, not necessarily other demographic or self-reported data.

EUR for European ancestry

CSA for Central/South Asian ancestry

AFR for African ancestry

EAS for East Asian ancestry

MID for Middle Eastern ancestry

AMR for Admixed American ancestry

outdir: Directory for which files will be saved.

start_from: Int specifying which file to start downloading at

num_files: Int specifying how many files to download

Multiple download sessions at once

Since there are lots of files to download, one can specify the number of files to download by modifying the num_files files option. This function will download the first num_files file counting from the start_from element of the output of get_ukb_filenames(). For example, if myfiles = get_ukb_filenames("afr"), then the function will download files from myfiles[start_from] through myfiles[start_from+num_files-1]

EasyLD.download_gnomad_variant_index_tables — Function

download_gnomad_variant_index_tables(population::String, outdir::String)

Downloads variant index hail tables from gnomAD panel for specified population and saves result to outdir. population accepts the same population labels as download_gnomad_LD_matrices. These files are typically at most a few GB.

EasyLD.download_ukb_variant_index_tables — Function

download_ukb_variant_index_tables(population::String, outdir::String)

Downloads variant index hail tables from pan-UK-biobank for specified population and saves result to outdir. population accepts the same population labels as download_ukb_LD_matrices. These files are typically at most a few GB.

EasyLD.get_gnomad_filenames — Function

get_gnomad_filenames(population::String)

Returns a list of file names that would be downloaded for the specified population. This function will not download anything. If join=true, the full path will be returned

EasyLD.get_ukb_filenames — Function

get_ukb_filenames(population::String)

Returns a list of file names that would be downloaded for the specified population. This function will not download anything. If join=true, the full path will be returned

EasyLD.hail_block_matrix — Function

hail_block_matrix(bm_files::String)

Creates a HailBlockMatrix which allows reading chunks of LD matrix data into memory. It is a subtype of AbstractMatrix, so operations like indexing and size() works, but only accessing its elements in contiguous chunks is "fast". I may support more functionalities depending on interest.

Examples

using EasyLD
data = "/Users/biona001/.julia/dev/EasyLD/data/gnomad.genomes.r2.1.1.nfe.common.adj.ld.bm"
bm = hail_block_matrix(data); # need a ';' to avoid displaying a few entries of bm, which takes ~0.1 seconds per entry

# get matrix dimension
size(bm) # returns (14207204, 14207204)

# read first 10k by 10k block into memory (takes roughly 7 seconds)
bm[1:10000, 1:10000]

# arbitrary slicing works but is very slow
bm[1:3, 1:2:100] # ~22 seconds

EasyLD.read_variant_index_tables — Function

read_variant_index_tables(ht_file::String, outdir::String)

Read variant index hail tables into a DataFrame. The first time this function gets called, we will read the original .ht files into memory and write the result to a tab-separated value .tsv file into the same directory as ht_file